PostgreSqlSegmentStore: use solution sets
Update storeSolution to reflect changes in solution schema. Now creates
a new solution entry for each set of segments stored as a solution and
sets this solution as taking precedence on completion. Requires two
queries since the ON INSERT RULE for solution precedence prevents it
being involved in a CTE, but performance is unchanged.